[WikiGames] entry dialog (#5252)
* - updates otd entry dialog to new design
- updates preference for otd entry dialog
- removes unused code
- adds maybeShowOnThisDayGameDialog function to baseActivity

* - adds strings resource and re-arranges the strings order in the strings.xml and qq/strings.xml

* - string fix
