+++ app/src/main/java/org/wikipedia/games/onthisday/OnThisDayGameViewModel.kt
@@ -239,18 +239,6 @@ class OnThisDayGameViewModel(savedStateHandle: SavedStateHandle) : ViewModel() {
         const val MAX_QUESTIONS = 10
         const val EXTRA_DATE = "date"
 
-        fun shouldShowEntryDialog(): Boolean {
-            if (Prefs.lastOtdGameVisitDate.isEmpty()) {
-                return true
-            }
-            try {
-                val prevDate = LocalDate.parse(Prefs.lastOtdGameVisitDate, DateTimeFormatter.ISO_LOCAL_DATE)
-                return prevDate.dayOfMonth != LocalDate.now().dayOfMonth
-            } catch (e: Exception) {
-                return true
-            }
-        }
-
         val gameStartDate: LocalDate get() {
             return try {
                 LocalDate.parse(Prefs.otdGameStartDate)
