+++ app/src/main/java/org/wikipedia/page/PageFragment.kt
@@ -18,8 +18,6 @@ import android.view.ViewGroup
 import android.webkit.WebResourceRequest
 import android.webkit.WebResourceResponse
 import android.webkit.WebView
-import android.widget.Button
-import android.widget.ImageView
 import android.widget.LinearLayout
 import androidx.appcompat.app.AppCompatActivity
 import androidx.coordinatorlayout.widget.CoordinatorLayout
@@ -47,6 +45,7 @@ import org.wikipedia.Constants.InvokeSource
 import org.wikipedia.LongPressHandler
 import org.wikipedia.R
 import org.wikipedia.WikipediaApp
+import org.wikipedia.activity.BaseActivity
 import org.wikipedia.activity.FragmentUtil.getCallback
 import org.wikipedia.analytics.eventplatform.ArticleFindInPageInteractionEvent
 import org.wikipedia.analytics.eventplatform.ArticleInteractionEvent
@@ -75,8 +74,6 @@ import org.wikipedia.descriptions.DescriptionEditActivity
 import org.wikipedia.diff.ArticleEditDetailsActivity
 import org.wikipedia.edit.EditHandler
 import org.wikipedia.gallery.GalleryActivity
-import org.wikipedia.games.onthisday.OnThisDayGameActivity
-import org.wikipedia.games.onthisday.OnThisDayGameViewModel
 import org.wikipedia.history.HistoryEntry
 import org.wikipedia.json.JsonUtil
 import org.wikipedia.login.LoginActivity
@@ -695,22 +692,6 @@ class PageFragment : Fragment(), BackPressedHandler, CommunicationBridge.Communi
         }
     }
 
-    private fun maybeShowOnThisDayGameDialog() {
-        if (Prefs.isOtdGameDialogEnabled && OnThisDayGameViewModel.shouldShowEntryDialog()) {
-            val dialogView = layoutInflater.inflate(R.layout.dialog_on_this_day_game, null)
-            val dialog = MaterialAlertDialogBuilder(requireActivity())
-                .setView(dialogView)
-                .show()
-            dialogView.findViewById<Button>(R.id.playGameButton).setOnClickListener {
-                startActivity(OnThisDayGameActivity.newIntent(requireContext(), InvokeSource.PAGE_ACTIVITY))
-                dialog.dismiss()
-            }
-            dialogView.findViewById<ImageView>(R.id.closeButton).setOnClickListener {
-                dialog.dismiss()
-            }
-        }
-    }
-
     private fun showFindReferenceInPage(referenceAnchor: String,
                                         backLinksList: List<String?>,
                                         referenceText: String) {
@@ -956,6 +937,10 @@ class PageFragment : Fragment(), BackPressedHandler, CommunicationBridge.Communi
         // handler), since the page metadata might have altered the lead image display state.
         bridge.execute(JavaScriptActionHandler.setTopMargin(leadImagesHandler.topMargin))
         bridge.execute(JavaScriptActionHandler.setFooter(model))
+    }
+
+    private fun maybeShowOnThisDayGameDialog() {
+        (requireActivity() as? BaseActivity)?.maybeShowOnThisDayGameDialog()
     }
 
     fun openInNewBackgroundTab(title: PageTitle, entry: HistoryEntry) {
