+++ app/src/main/res/values/preference_keys.xml
@@ -171,11 +171,11 @@
     <string name="preference_key_payment_methods_merchant_id">paymentMethodsMerchantId</string>
     <string name="preference_key_payment_methods_gateway_id">paymentMethodsGatewayId</string>
     <string name="preference_key_otd_game_state">otdGameState</string>
-    <string name="preference_key_otd_game_dialog_enable">otdGameDialogEnabled</string>
     <string name="preference_key_otd_game_last_visit_date">otdGameLastVisitDate</string>
     <string name="preference_key_otd_game_start_date">otdGameStartDate</string>
     <string name="preference_key_otd_game_end_date">otdGameEndDate</string>
     <string name="preference_key_otd_game_num_questions">otdGameNumQuestions</string>
+    <string name="preference_key_otd_entry_dialog_shown">otdEntryDialogShown</string>
     <string name="preference_key_default_places_location_latlng">placesDefaultLocationLatLng</string>
     <string name="preference_key_delete_local_donation_history">deleteLocalDonationHistory</string>
 </resources>
