+++ app/src/main/res/xml/developer_preferences.xml
@@ -457,8 +457,8 @@
     <PreferenceCategory android:title="On this day game">
 
         <org.wikipedia.settings.SwitchPreferenceMultiLine
-            android:key="@string/preference_key_otd_game_dialog_enable"
-            android:title="@string/preference_key_otd_game_dialog_enable" />
+            android:key="@string/preference_key_otd_entry_dialog_shown"
+            android:title="@string/preference_key_otd_entry_dialog_shown" />
 
         <org.wikipedia.settings.EditTextAutoSummarizePreference
             style="@style/EditTextAutoSummarizePreference"
